Sort games by match date/time, and SVG icons, and refactor a little.
Save the datetime from the match info and sort games by the match time,
since the datetime for the game info provided by the API is unreliable.

Move the code that generates the HTML for each game info into
generateGameHtml().

Replace all HTML with Html for consistency.

Add icons.svg file containing icons for kills and gold, and insert it at
the beginning of the document so the icons can be used on the page.

Add a CSS file.

Make html() accept empty arrays (i.e. html([]) now returns '') and
reject undefined.