Lots of changes.
Remove app/ folder.

Create public/ folder for generated HTML/JavaScript.

Create server-render.ts, which renders the app on the server side and
saves the HTML to public/index.html, but also generates a JavaScript
bundle that includes the data so that the React components will "come
alive" and app can be used interactively when the JavaScript finishes
loading.

Organize code into server/ folder to hold the serer side rendering and
game data fetching code, client/ to hold the code that will be run in
the browser, and shared/ for code used by both (such as the React
components).