Separate JavaScript bundle and game data, and create simple logging.
Instead of including the game data generated from the lolesports API
directly in the bundle for the client side JavaScript, insert the data
directly into the generated HTML so that the JavaScript bundle can be
cached by the user's browser (since the bundle includes React and all of
its dependencies, it can be pretty large, so not having to reload it
each time new games are added to the list should make a big difference).

Also, create util.log() and util.error() functions so that I can capture
all of the log/error messages and email them to myself as a very simple
form of monitoring.

Finally, separate constants into a separate shared/constants.ts file.